- Added reference counting to pipes. - Made pipe implementation private.
