+++ kernel/pipe.h
@@ -8,26 +8,16 @@
 #define PIPE_H
 
 #include "kernel/types.h"
-#include "list.h"
 
-#define PIPE_SIZE (1024)
-
-struct pipe {
-	char *buffer;
-	int read_pos;
-	int write_pos;
-	int blocking;
-	int flushed;
-	struct list queue;
-};
-
-struct pipe *pipe_open();
-void pipe_close(struct pipe *p);
+struct pipe *pipe_create();
+struct pipe *pipe_addref( struct pipe *p );
+void pipe_delete(struct pipe *p);
 void pipe_flush(struct pipe *p);
 int pipe_set_blocking(struct pipe *p, int b);
 
 int pipe_write(struct pipe *p, char *buffer, int size);
 int pipe_read(struct pipe *p, char *buffer, int size);
 int pipe_read_nonblock(struct pipe *p, char *buffer, int size);
+int pipe_size( struct pipe *p);
 
 #endif
